+++ crates/config/src/renderers/typescript.rs
@@ -1,6 +1,6 @@
 use crate::schema::{RenderResult, SchemaRenderer};
 use schematic_types::*;
-use std::collections::HashSet;
+use std::collections::{HashMap, HashSet};
 
 /// Format of a TypeScript enum.
 #[derive(Default)]
@@ -27,14 +27,33 @@ pub enum ObjectFormat {
 /// Options to control the rendered TypeScript output.
 #[derive(Default)]
 pub struct TypeScriptOptions {
+    /// Render a `const enum` instead of a `enum`.
     pub const_enum: bool,
+
+    /// Disable references and render all types inline recursively.
+    pub disable_references: bool,
+
+    /// Format to render enums, either an `enum` or a `type` union.
     pub enum_format: EnumFormat,
+
+    /// List of references to exclude from exporting as a type.
+    pub exclude_references: HashSet<String>,
+
+    /// Map of relative import file paths to a list of types to import.
+    /// Will be rendered as an `import type {} from 'path';` declaration.
+    pub external_types: HashMap<String, HashSet<String>>,
+
+    /// Character(s) to use for indentation.
+    pub indent_char: String,
+
+    /// Format to render objects, either an `interface` or `type`.
     pub object_format: ObjectFormat,
 }
 
 /// Renders TypeScript types from a schema.
 #[derive(Default)]
 pub struct TypeScriptRenderer {
+    depth: usize,
     options: TypeScriptOptions,
     references: HashSet<String>,
 }
@@ -42,27 +61,58 @@ pub struct TypeScriptRenderer {
 impl TypeScriptRenderer {
     pub fn new(options: TypeScriptOptions) -> Self {
         Self {
+            depth: 0,
             options,
             references: HashSet::new(),
         }
     }
 
-    fn is_string_union_enum(&self, enu: &EnumType) -> bool {
-        matches!(self.options.enum_format, EnumFormat::Union) || enu.variants.is_none()
+    fn indent(&self) -> String {
+        let chars = if self.options.indent_char.is_empty() {
+            "\t"
+        } else {
+            &self.options.indent_char
+        };
+
+        if self.depth == 0 {
+            String::new()
+        } else {
+            chars.repeat(self.depth)
+        }
     }
 
-    fn export_type_alias(&self, name: &str, value: String) -> RenderResult {
+    fn is_excluded(&self, name: &str) -> bool {
+        self.options.exclude_references.contains(name)
+    }
+
+    fn is_external(&self, name: &str) -> bool {
+        for externals in self.options.external_types.values() {
+            if externals.contains(name) {
+                return true;
+            }
+        }
+
+        false
+    }
+
+    fn is_string_union_enum(&self, enu: &EnumType) -> bool {
+        matches!(self.options.enum_format, EnumFormat::Union)
+            || enu.variants.is_none()
+            || self.options.disable_references
+    }
+
+    fn export_type_alias(&mut self, name: &str, value: String) -> RenderResult {
         Ok(format!("export type {} = {};", name, value))
     }
 
-    fn export_enum_type(&self, name: &str, enu: &EnumType) -> RenderResult {
+    fn export_enum_type(&mut self, name: &str, enu: &EnumType) -> RenderResult {
         let value = self.render_enum(enu)?;
 
         if self.is_string_union_enum(enu) {
             return self.export_type_alias(name, value);
         }
 
-        let out = format!("enum {} {{\n{}\n}}", name, value);
+        let out = format!("enum {} {}", name, value);
 
         Ok(if self.options.const_enum {
             format!("export const {}", out)
@@ -71,7 +121,7 @@ impl TypeScriptRenderer {
         })
     }
 
-    fn export_object_type(&self, name: &str, structure: &StructType) -> RenderResult {
+    fn export_object_type(&mut self, name: &str, structure: &StructType) -> RenderResult {
         let value = self.render_struct(structure)?;
 
         if matches!(self.options.object_format, ObjectFormat::Interface) {
@@ -81,7 +131,7 @@ impl TypeScriptRenderer {
         self.export_type_alias(name, value)
     }
 
-    fn render_enum_or_union(&self, enu: &EnumType) -> RenderResult {
+    fn render_enum_or_union(&mut self, enu: &EnumType) -> RenderResult {
         if self.is_string_union_enum(enu) {
             return self.render_union(&UnionType {
                 variants_types: enu
@@ -94,7 +144,10 @@ impl TypeScriptRenderer {
             });
         }
 
-        let mut fields = vec![];
+        self.depth += 1;
+
+        let mut out = vec![];
+        let indent = self.indent();
 
         for variant in enu.variants.as_ref().unwrap() {
             if variant.hidden {
@@ -104,68 +157,87 @@ impl TypeScriptRenderer {
             if let Some(variant_name) = &variant.name {
                 let mut field = if matches!(self.options.enum_format, EnumFormat::ValuedEnum) {
                     format!(
-                        "\t{} = {},",
+                        "{}{} = {},",
+                        indent,
                         variant_name,
                         self.render_schema(&variant.type_of)?
                     )
                 } else {
-                    format!("\t{},", variant_name)
+                    format!("{}{},", indent, variant_name)
                 };
 
                 if let Some(comment) = &variant.description {
                     field = self.wrap_in_comment(comment.trim(), field);
                 }
 
-                fields.push(field);
+                out.push(field);
             }
         }
 
-        Ok(fields.join("\n"))
+        self.depth -= 1;
+
+        Ok(format!("{{\n{}\n{}}}", out.join("\n"), self.indent()))
     }
 
     fn wrap_in_comment(&self, comment: &str, value: String) -> String {
+        let indent = self.indent();
+
         if comment.starts_with('*') {
-            let mut out = vec!["\t/**".to_owned()];
+            let mut out = vec![format!("{}/**", indent)];
 
             for line in comment.split('\n') {
-                out.push(format!("\t {}", line.trim()));
+                out.push(format!("{} {}", indent, line.trim()));
             }
 
-            out.push("\t */".to_owned());
+            out.push(format!("{} */", indent));
 
             format!("{}\n{}", out.join("\n"), value)
         } else {
-            format!("\t// {}\n{}", comment.trim(), value)
+            format!("{}// {}\n{}", indent, comment.trim(), value)
         }
     }
 }
 
 impl SchemaRenderer<String> for TypeScriptRenderer {
     fn is_reference(&self, name: &str) -> bool {
-        self.references.contains(name)
+        if self.options.disable_references {
+            return false;
+        }
+
+        if self.references.contains(name) {
+            return true;
+        }
+
+        self.is_external(name)
     }
 
-    fn render_array(&self, array: &ArrayType) -> RenderResult {
-        Ok(format!("{}[]", self.render_schema(&array.items_type)?))
+    fn render_array(&mut self, array: &ArrayType) -> RenderResult {
+        let out = self.render_schema(&array.items_type)?;
+
+        Ok(if out.contains('|') {
+            format!("({})[]", out)
+        } else {
+            format!("{}[]", out)
+        })
     }
 
-    fn render_boolean(&self) -> RenderResult {
+    fn render_boolean(&mut self) -> RenderResult {
         Ok("boolean".into())
     }
 
-    fn render_enum(&self, enu: &EnumType) -> RenderResult {
+    fn render_enum(&mut self, enu: &EnumType) -> RenderResult {
         self.render_enum_or_union(enu)
     }
 
-    fn render_float(&self, _: &FloatType) -> RenderResult {
+    fn render_float(&mut self, _: &FloatType) -> RenderResult {
         Ok("number".into())
     }
 
-    fn render_integer(&self, _: &IntegerType) -> RenderResult {
+    fn render_integer(&mut self, _: &IntegerType) -> RenderResult {
         Ok("number".into())
     }
 
-    fn render_literal(&self, literal: &LiteralType) -> RenderResult {
+    fn render_literal(&mut self, literal: &LiteralType) -> RenderResult {
         if let Some(value) = &literal.value {
             return Ok(match value {
                 LiteralValue::Bool(inner) => inner.to_string(),
@@ -179,11 +251,11 @@ impl SchemaRenderer<String> for TypeScriptRenderer {
         self.render_unknown()
     }
 
-    fn render_null(&self) -> RenderResult {
+    fn render_null(&mut self) -> RenderResult {
         Ok("null".into())
     }
 
-    fn render_object(&self, object: &ObjectType) -> RenderResult {
+    fn render_object(&mut self, object: &ObjectType) -> RenderResult {
         Ok(format!(
             "Record<{}, {}>",
             self.render_schema(&object.key_type)?,
@@ -191,23 +263,26 @@ impl SchemaRenderer<String> for TypeScriptRenderer {
         ))
     }
 
-    fn render_reference(&self, reference: &str) -> RenderResult {
+    fn render_reference(&mut self, reference: &str) -> RenderResult {
         Ok(reference.into())
     }
 
-    fn render_string(&self, _: &StringType) -> RenderResult {
+    fn render_string(&mut self, _: &StringType) -> RenderResult {
         Ok("string".into())
     }
 
-    fn render_struct(&self, structure: &StructType) -> RenderResult {
+    fn render_struct(&mut self, structure: &StructType) -> RenderResult {
+        self.depth += 1;
+
         let mut out = vec![];
+        let indent = self.indent();
 
         for field in &structure.fields {
             if field.hidden {
                 continue;
             }
 
-            let mut row = format!("\t{}", field.name.as_ref().unwrap());
+            let mut row = format!("{}{}", indent, field.name.as_ref().unwrap());
 
             if field.optional {
                 row.push_str("?: ");
@@ -234,10 +309,12 @@ impl SchemaRenderer<String> for TypeScriptRenderer {
             out.push(row);
         }
 
-        Ok(format!("{{\n{}\n}}", out.join("\n")))
+        self.depth -= 1;
+
+        Ok(format!("{{\n{}\n{}}}", out.join("\n"), self.indent()))
     }
 
-    fn render_tuple(&self, tuple: &TupleType) -> RenderResult {
+    fn render_tuple(&mut self, tuple: &TupleType) -> RenderResult {
         let mut items = vec![];
 
         for item in &tuple.items_types {
@@ -247,7 +324,7 @@ impl SchemaRenderer<String> for TypeScriptRenderer {
         Ok(format!("[{}]", items.join(", ")))
     }
 
-    fn render_union(&self, uni: &UnionType) -> RenderResult {
+    fn render_union(&mut self, uni: &UnionType) -> RenderResult {
         let mut items = vec![];
 
         for item in &uni.variants_types {
@@ -257,7 +334,7 @@ impl SchemaRenderer<String> for TypeScriptRenderer {
         Ok(items.join(" | "))
     }
 
-    fn render_unknown(&self) -> RenderResult {
+    fn render_unknown(&mut self) -> RenderResult {
         Ok("unknown".into())
     }
 
@@ -269,13 +346,35 @@ impl SchemaRenderer<String> for TypeScriptRenderer {
             "/* eslint-disable */".to_string(),
         ];
 
+        let mut imports = vec![];
+
+        for (import, types) in &self.options.external_types {
+            let mut imported_types = types.iter().cloned().collect::<Vec<_>>();
+            imported_types.sort();
+
+            imports.push(format!(
+                "import type {{ {} }} from '{}';",
+                imported_types.join(", "),
+                import
+            ));
+        }
+
+        if !imports.is_empty() {
+            outputs.push(imports.join("\n"));
+        }
+
         for schema in schemas {
             if let Some(name) = schema.get_name() {
+                if self.is_excluded(name) {
+                    continue;
+                }
+
                 outputs.push(match schema {
                     SchemaType::Enum(inner) => self.export_enum_type(name, inner)?,
                     SchemaType::Struct(inner) => self.export_object_type(name, inner)?,
                     _ => {
-                        self.export_type_alias(name, self.render_schema_without_reference(schema)?)?
+                        let out = self.render_schema_without_reference(schema)?;
+                        self.export_type_alias(name, out)?
                     }
                 });
             }